+++ src/HookRegistry.php
@@ -63,13 +63,18 @@ class HookRegistry {
 
 	private function addCallbackHandlers( $store, $options ) {
 
-		$propertyRegistry = new PropertyRegistry();
-
 		/**
 		 * @see https://github.com/SemanticMediaWiki/SemanticMediaWiki/blob/master/docs/technical/hooks.md
 		 */
-		$this->handlers['SMW::Property::initProperties'] = function () use ( $propertyRegistry ) {
-			return $propertyRegistry->register();
+		$this->handlers['SMW::Property::initProperties'] = function( $baseRegistry ) {
+
+			$propertyRegistry = new PropertyRegistry();
+
+			$propertyRegistry->register(
+				$baseRegistry
+			);
+
+			return true;
 		};
 
 		/**