+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@ This file contains the RELEASE-NOTES of the Semantic Breadcrumb Links (a.k.a. SB
 
 ### 1.3.0 (dev)
 
+* Support for `Display title of` provided by SMW 2.4
 * Added `$GLOBALS['egSBLEnabledSubpageParentAnnotation']` to allow for subpage
   parents to be annotated automatically if no other `Has parent page` exists
 * Recognize `egSBLPageTitleToHideSubpageParent` / `wgNamespacesWithSubpages`